linked food diary and food logging with luca, all checks passed. Had to move a few files around to pass the detekt. created helper files to account for this
