+++ app/src/main/kotlin/diettracker/CaloriesSession.kt
@@ -3,8 +3,12 @@ package diettracker
 import diettracker.db.tables.Foods
 import diettracker.db.tables.RecipeIngredients
 import diettracker.db.tables.Recipes
+import diettracker.models.CurrentMealFood
+import diettracker.models.CurrentMealSession
 import diettracker.models.Food
 import diettracker.models.Recipe
+import diettracker.services.DiaryService.getMealTypeByTime
+import diettracker.services.DiaryService.saveFoodLog
 import io.ktor.server.application.ApplicationCall
 import io.ktor.server.pebble.respondTemplate
 import io.ktor.server.request.receiveParameters
@@ -62,7 +66,7 @@ suspend fun ApplicationCall.foodLogRecipe() {
         )
         return
     }
-
+    val foodsToAdd = mutableListOf<CurrentMealFood>()
     transaction {
         val ingredients =
             RecipeIngredients
@@ -81,6 +85,8 @@ suspend fun ApplicationCall.foodLogRecipe() {
             addProtein += protein
             addFat += fat
             addCarbs += carbs
+
+            foodsToAdd.add(CurrentMealFood(foodId, grams))
         }
     }
 
@@ -91,6 +97,12 @@ suspend fun ApplicationCall.foodLogRecipe() {
     val newTotalCarbs = caloriesSession.carbs + addCarbs
 
     sessions.set(CaloriesSession(newTotalCals, newTotalProtein, newTotalFat, newTotalCarbs))
+    val currentMeal = sessions.get<CurrentMealSession>() ?: CurrentMealSession(emptyList())
+    sessions.set(
+        CurrentMealSession(
+            currentMeal.foods + foodsToAdd,
+        ),
+    )
     respondRedirect("/food_log")
 }
 
@@ -135,6 +147,20 @@ suspend fun ApplicationCall.foodLogCustom() {
 
     sessions.set(CaloriesSession(newTotalCals, newTotalProtein, newTotalFat, newTotalCarbs))
 
+    val currentMeal = sessions.get<CurrentMealSession>() ?: CurrentMealSession(emptyList())
+
+    val updatedMeal =
+        currentMeal.copy(
+            foods =
+                currentMeal.foods +
+                    CurrentMealFood(
+                        foodId = foodId,
+                        grams = grams,
+                    ),
+        )
+
+    sessions.set(CurrentMealSession(updatedMeal.foods))
+
     respondTemplate(
         "pages/client_dash/add_food.peb",
         mapOf(
@@ -150,11 +176,7 @@ fun searchRecipes(query: String): List<Recipe> =
     transaction {
         val searchTerm = query.lowercase()
 
-        if (searchTerm.isBlank()) {
-            return@transaction emptyList<Recipe>()
-        }
-
-        if (searchTerm.any { it.isDigit() }) {
+        if (searchTerm.isBlank() || searchTerm.any { it.isDigit() }) {
             return@transaction emptyList<Recipe>()
         }
 
@@ -176,11 +198,7 @@ fun searchFoods(foodquery: String): List<Food> =
     transaction {
         val searchTerm = foodquery.lowercase()
 
-        if (searchTerm.isBlank()) {
-            return@transaction emptyList<Food>()
-        }
-
-        if (searchTerm.any { it.isDigit() }) {
+        if (searchTerm.isBlank() || searchTerm.any { it.isDigit() }) {
             return@transaction emptyList<Food>()
         }
 
@@ -202,66 +220,6 @@ fun searchFoods(foodquery: String): List<Food> =
         return@transaction foods
     }
 
-fun calcCalcsById(
-    foodid: Int,
-    grams: Int,
-): Int {
-    val multiplier = grams / GRAMS_PER_SERVING.toDouble()
-    val caloriesPer100g =
-        Foods
-            .selectAll()
-            .where { Foods.food_id eq foodid }
-            .map { row -> row[Foods.calories_per_100g].toDouble().toInt() }
-            .firstOrNull() ?: return 0
-
-    return (caloriesPer100g * multiplier).toInt()
-}
-
-fun calcProteinById(
-    foodid: Int,
-    grams: Int,
-): Int {
-    val multiplier = grams / GRAMS_PER_SERVING.toDouble()
-    val proteinPer100g =
-        Foods
-            .selectAll()
-            .where { Foods.food_id eq foodid }
-            .map { row -> row[Foods.protein_per_100g].toDouble().toInt() }
-            .firstOrNull() ?: return 0
-
-    return (proteinPer100g * multiplier).toInt()
-}
-
-fun calcFatById(
-    foodid: Int,
-    grams: Int,
-): Int {
-    val multiplier = grams / GRAMS_PER_SERVING.toDouble()
-    val fatPer100g =
-        Foods
-            .selectAll()
-            .where { Foods.food_id eq foodid }
-            .map { row -> row[Foods.fat_per_100g].toDouble().toInt() }
-            .firstOrNull() ?: return 0
-
-    return (fatPer100g * multiplier).toInt()
-}
-
-fun calcCarbsById(
-    foodid: Int,
-    grams: Int,
-): Int {
-    val multiplier = grams / GRAMS_PER_SERVING.toDouble()
-    val carbsPer100g =
-        Foods
-            .selectAll()
-            .where { Foods.food_id eq foodid }
-            .map { row -> row[Foods.carbs_per_100g].toDouble().toInt() }
-            .firstOrNull() ?: return 0
-
-    return (carbsPer100g * multiplier).toInt()
-}
-
 suspend fun ApplicationCall.foodLogReset() {
     sessions.set(CaloriesSession(0, 0, 0, 0))
     respondTemplate(
@@ -273,4 +231,30 @@ suspend fun ApplicationCall.foodLogReset() {
             "carbs" to 0,
         ),
     )
+}
+
+suspend fun ApplicationCall.saveCurrentFoodLog() {
+    val mealType = getMealTypeByTime()
+    val notes = ""
+    val userSession = sessions.get<UserSession>()
+    val userId = userSession?.let { getUserIdByEmail(it.email) }
+    val currentMealSession = sessions.get<CurrentMealSession>()
+
+    when {
+        userId == null -> respondRedirect("/login")
+        currentMealSession == null || currentMealSession.foods.isEmpty() -> respondRedirect("/food_log")
+        else -> {
+            saveFoodLog(
+                userId = userId,
+                mealType = mealType,
+                notes = notes,
+                foods = currentMealSession.foods,
+            )
+
+            sessions.set(CaloriesSession(0, 0, 0, 0))
+            sessions.set(CurrentMealSession(emptyList()))
+
+            respondRedirect("/food_log")
+        }
+    }
 }
